rename main to greeting and del sanya
